Update draw.lua
Especially for systems that do not separate telemetry.statusArmed and telemetry.landComplete, but handle them together (Mav2PT, OlliW MavSDK), the previous code did not display "ARMED" dialog correctly. If it was seen at all was dependent on the (random) state of timer for blinkon variable at the time of call to drawArmStatus().
The change here defaults to approx. 1.5 second display when state is changed from disarmed to armed, which can be shortened only in case in this time disarmed is again chosen. The result is clearly displayed "ARMED" message.