Added an RSSI symbol (I drew a simple antenna). I did not touch the CSRF RSSI display though as don't have hw to test it.
Additionally, this commit wraps all mavsdk calls into
if mavsdk ~= nil then ... end
wrapper, so that on non-OlliW OpenTX firmwares the script does not get disabled anymore even if the user selects from configuration to enable MAVLink nevertheless.
I might add a warning pop-up of some sorts as a feedback for the user of this misconfiguration.